Fix(event-broker): Extract test-util-only helpers into dedicated module
Move dispatcher test helpers used exclusively by #[cfg(feature = "test-util")]
tests into a separate dispatcher_test_util module, gated once in mod.rs.

Fixes clippy dead_code errors when building with --no-default-features
--all-targets.
