+++ gears/system/event-broker/event-broker-sdk/src/consumer/dispatcher_tests.rs
@@ -1,5 +1,3 @@
-#[cfg(feature = "test-util")]
-use std::collections::BTreeSet;
 use std::collections::HashMap;
 use std::sync::{Arc, Mutex};
 use std::time::Duration;
@@ -18,36 +16,25 @@ use super::{
     BatchHandlerOutcome, ConsumerHandler, EventBatch, EventTypeRef, HandlerOutcome, RawEvent,
     SingleEventHandlerAdapter, TopicRef,
 };
-#[cfg(feature = "test-util")]
-use super::{
-    CommitOffset, ConnectionDropReason, ConsumerBuffering, ConsumerBuilder, ConsumerCommitMode,
-    ConsumerGroupRef, ConsumerListenerSettings, ConsumerRuntimeEvent, ConsumerRuntimeListener,
-    ConsumerSlowDetection, Fallback, InMemoryOffsetManager, OffsetManagerError, OffsetStore,
-    PartitionBufferState, ResolvedPosition, SlowConsumerTrigger,
-};
 use crate::error::ConsumerError;
-#[cfg(feature = "test-util")]
-use crate::ids::ConsumerGroupId;
 use crate::ids::TopicId;
+#[cfg(feature = "test-util")]
+use {
+    super::dispatcher_test_util::*,
+    super::{
+        CommitOffset, ConnectionDropReason, ConsumerBuffering, ConsumerBuilder, ConsumerCommitMode,
+        ConsumerGroupRef, ConsumerListenerSettings, ConsumerRuntimeEvent, ConsumerSlowDetection,
+        Fallback, InMemoryOffsetManager, OffsetManagerError, OffsetStore, PartitionBufferState,
+        ResolvedPosition, SlowConsumerTrigger,
+    },
+    crate::ids::ConsumerGroupId,
+    std::collections::BTreeSet,
+};
 
 type SharedOffsets = Arc<Mutex<Vec<i64>>>;
 type SharedNamedOffsets = Arc<Mutex<Vec<(&'static str, i64)>>>;
 type PartitionCall = (String, u32, i64);
 type SharedPartitionCalls = Arc<Mutex<Vec<PartitionCall>>>;
-#[cfg(feature = "test-util")]
-type SharedAttempts = Arc<Mutex<Vec<u16>>>;
-#[cfg(feature = "test-util")]
-type SharedTimeline = Arc<Mutex<Vec<&'static str>>>;
-#[cfg(feature = "test-util")]
-type CommitRecord = (ConsumerGroupId, TopicId, u32, i64);
-#[cfg(feature = "test-util")]
-type SharedCommits = Arc<Mutex<Vec<CommitRecord>>>;
-#[cfg(feature = "test-util")]
-type SharedScopes = Arc<Mutex<Vec<(String, u32, usize)>>>;
-#[cfg(feature = "test-util")]
-type SharedViolations = Arc<Mutex<Vec<String>>>;
-#[cfg(feature = "test-util")]
-type SharedRuntimeEvents = Arc<Mutex<Vec<ConsumerRuntimeEvent>>>;
 type TestPartitionBuffers =
     Arc<RwLock<HashMap<TopicPartitionKey, super::dispatcher::PartitionEventBuffer>>>;
 
@@ -72,17 +59,6 @@ fn raw_event_on(topic: &str, type_id: &str, partition: u32, offset: i64) -> RawE
         trace_parent: None,
         data: serde_json::json!({ "offset": offset }),
     }
-}
-
-#[cfg(feature = "test-util")]
-fn partition_key_for_partition(target: u32, partitions: u32) -> String {
-    assert_eq!(partitions, 2, "only the two-partition fixture is supported");
-    match target {
-        0 => "partition-key-0-1",
-        1 => "partition-key-1-0",
-        _ => panic!("two-partition fixture cannot target partition {target}"),
-    }
-    .to_owned()
 }
 
 struct RecordingSingleHandler {
@@ -139,284 +115,6 @@ impl ConsumerHandler for AckAllBatchHandler {
                 .iter()
                 .map(|event| (event.topic.clone(), event.partition, event.offset)),
         );
-        Ok(chunk
-            .last()
-            .map(|event| BatchHandlerOutcome::AdvanceThrough {
-                offset: event.offset,
-            })
-            .unwrap_or(BatchHandlerOutcome::Success))
-    }
-}
-
-#[cfg(feature = "test-util")]
-struct SleepingBatchHandler {
-    calls: SharedPartitionCalls,
-    delay: Duration,
-}
-
-#[cfg(feature = "test-util")]
-#[async_trait::async_trait]
-impl ConsumerHandler for SleepingBatchHandler {
-    async fn handle_batch(
-        &self,
-        batch: &EventBatch<'_>,
-        _attempts: u16,
-    ) -> Result<BatchHandlerOutcome, ConsumerError> {
-        tokio::time::sleep(self.delay).await;
-        let chunk = batch.next_chunk(batch.len());
-        self.calls.lock().unwrap().extend(
-            chunk
-                .iter()
-                .map(|event| (event.topic.clone(), event.partition, event.offset)),
-        );
-        Ok(chunk
-            .last()
-            .map(|event| BatchHandlerOutcome::AdvanceThrough {
-                offset: event.offset,
-            })
-            .unwrap_or(BatchHandlerOutcome::Success))
-    }
-}
-
-#[cfg(feature = "test-util")]
-struct FailingThenCommitBatchHandler {
-    failures_remaining: Arc<Mutex<usize>>,
-    calls: SharedAttempts,
-}
-
-#[cfg(feature = "test-util")]
-#[async_trait::async_trait]
-impl ConsumerHandler for FailingThenCommitBatchHandler {
-    async fn handle_batch(
-        &self,
-        _batch: &EventBatch<'_>,
-        attempts: u16,
-    ) -> Result<BatchHandlerOutcome, ConsumerError> {
-        {
-            let mut guard = self.failures_remaining.lock().unwrap();
-            if *guard > 0 {
-                *guard -= 1;
-                return Err(ConsumerError::Internal(
-                    "intentional representative handler failure".to_owned(),
-                ));
-            }
-        }
-
-        self.calls.lock().unwrap().push(attempts);
-        Ok(BatchHandlerOutcome::Success)
-    }
-}
-
-#[cfg(feature = "test-util")]
-struct SequencedOffsetManager {
-    timeline: SharedTimeline,
-}
-
-#[cfg(feature = "test-util")]
-#[async_trait::async_trait]
-impl OffsetStore for SequencedOffsetManager {
-    async fn load_position(
-        &self,
-        _group: &ConsumerGroupId,
-        _topic: &TopicId,
-        _partition: u32,
-    ) -> Result<ResolvedPosition, OffsetManagerError> {
-        self.timeline.lock().unwrap().push("load");
-        Ok(ResolvedPosition::Earliest)
-    }
-}
-
-#[cfg(feature = "test-util")]
-#[async_trait::async_trait]
-impl CommitOffset for SequencedOffsetManager {
-    async fn commit(
-        &self,
-        _group: &ConsumerGroupId,
-        _topic: &TopicId,
-        _partition: u32,
-        _offset: i64,
-    ) -> Result<(), OffsetManagerError> {
-        Ok(())
-    }
-}
-
-#[cfg(feature = "test-util")]
-#[derive(Clone, Default)]
-struct RecordingCommitOffsetManager {
-    commits: SharedCommits,
-}
-
-#[cfg(feature = "test-util")]
-#[async_trait::async_trait]
-impl OffsetStore for RecordingCommitOffsetManager {
-    async fn load_position(
-        &self,
-        _group: &ConsumerGroupId,
-        _topic: &TopicId,
-        _partition: u32,
-    ) -> Result<ResolvedPosition, OffsetManagerError> {
-        Ok(ResolvedPosition::Earliest)
-    }
-}
-
-#[cfg(feature = "test-util")]
-#[async_trait::async_trait]
-impl CommitOffset for RecordingCommitOffsetManager {
-    async fn commit(
-        &self,
-        group: &ConsumerGroupId,
-        topic: &TopicId,
-        partition: u32,
-        offset: i64,
-    ) -> Result<(), OffsetManagerError> {
-        self.commits
-            .lock()
-            .unwrap()
-            .push((*group, *topic, partition, offset));
-        Ok(())
-    }
-}
-
-#[cfg(feature = "test-util")]
-struct SequencedBatchHandler {
-    timeline: SharedTimeline,
-}
-
-#[cfg(feature = "test-util")]
-#[async_trait::async_trait]
-impl ConsumerHandler for SequencedBatchHandler {
-    async fn handle_batch(
-        &self,
-        _batch: &EventBatch<'_>,
-        _attempts: u16,
-    ) -> Result<BatchHandlerOutcome, ConsumerError> {
-        self.timeline.lock().unwrap().push("handle");
-        Ok(BatchHandlerOutcome::Success)
-    }
-}
-
-#[cfg(feature = "test-util")]
-#[derive(Default)]
-struct BatchScopeRecorder {
-    scopes: SharedScopes,
-    violations: SharedViolations,
-}
-
-#[cfg(feature = "test-util")]
-#[derive(Clone, Default)]
-struct RecordingRuntimeListener {
-    events: SharedRuntimeEvents,
-}
-
-#[cfg(feature = "test-util")]
-#[async_trait::async_trait]
-impl ConsumerRuntimeListener for RecordingRuntimeListener {
-    async fn on_consumer_event(&self, event: &ConsumerRuntimeEvent) -> Result<(), ConsumerError> {
-        self.events.lock().unwrap().push(event.clone());
-        Ok(())
-    }
-}
-
-#[cfg(feature = "test-util")]
-struct FailingRuntimeListener;
-
-#[cfg(feature = "test-util")]
-#[async_trait::async_trait]
-impl ConsumerRuntimeListener for FailingRuntimeListener {
-    async fn on_consumer_event(&self, _event: &ConsumerRuntimeEvent) -> Result<(), ConsumerError> {
-        Err(ConsumerError::Internal(
-            "intentional listener failure".to_owned(),
-        ))
-    }
-}
-
-#[cfg(feature = "test-util")]
-struct SlowRuntimeListener {
-    delay: Duration,
-}
-
-#[cfg(feature = "test-util")]
-#[async_trait::async_trait]
-impl ConsumerRuntimeListener for SlowRuntimeListener {
-    async fn on_consumer_event(&self, _event: &ConsumerRuntimeEvent) -> Result<(), ConsumerError> {
-        tokio::time::sleep(self.delay).await;
-        Ok(())
-    }
-}
-
-#[cfg(feature = "test-util")]
-#[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord)]
-enum RuntimeEventKind {
-    SubscriptionJoining,
-    SubscriptionStarted,
-    SubscriptionRejoining,
-    SubscriptionTerminated,
-    SubscriptionConnectionDropped,
-    AssignmentChanged,
-    ProgressAdvanced,
-    PartitionBufferStateChanged,
-    HandlerBatchStarted,
-    HandlerBatchCompleted,
-    HandlerFailed,
-    OffsetLoaded,
-    OffsetCommitted,
-    RetryScheduled,
-}
-
-#[cfg(feature = "test-util")]
-fn runtime_event_kind(event: &ConsumerRuntimeEvent) -> RuntimeEventKind {
-    match event {
-        ConsumerRuntimeEvent::SubscriptionJoining { .. } => RuntimeEventKind::SubscriptionJoining,
-        ConsumerRuntimeEvent::SubscriptionStarted { .. } => RuntimeEventKind::SubscriptionStarted,
-        ConsumerRuntimeEvent::SubscriptionRejoining { .. } => {
-            RuntimeEventKind::SubscriptionRejoining
-        }
-        ConsumerRuntimeEvent::SubscriptionTerminated { .. } => {
-            RuntimeEventKind::SubscriptionTerminated
-        }
-        ConsumerRuntimeEvent::SubscriptionConnectionDropped { .. } => {
-            RuntimeEventKind::SubscriptionConnectionDropped
-        }
-        ConsumerRuntimeEvent::AssignmentChanged { .. } => RuntimeEventKind::AssignmentChanged,
-        ConsumerRuntimeEvent::ProgressAdvanced { .. } => RuntimeEventKind::ProgressAdvanced,
-        ConsumerRuntimeEvent::PartitionBufferStateChanged { .. } => {
-            RuntimeEventKind::PartitionBufferStateChanged
-        }
-        ConsumerRuntimeEvent::HandlerBatchStarted { .. } => RuntimeEventKind::HandlerBatchStarted,
-        ConsumerRuntimeEvent::HandlerBatchCompleted { .. } => {
-            RuntimeEventKind::HandlerBatchCompleted
-        }
-        ConsumerRuntimeEvent::HandlerFailed { .. } => RuntimeEventKind::HandlerFailed,
-        ConsumerRuntimeEvent::OffsetLoaded { .. } => RuntimeEventKind::OffsetLoaded,
-        ConsumerRuntimeEvent::OffsetCommitted { .. } => RuntimeEventKind::OffsetCommitted,
-        ConsumerRuntimeEvent::RetryScheduled { .. } => RuntimeEventKind::RetryScheduled,
-    }
-}
-
-#[cfg(feature = "test-util")]
-#[async_trait::async_trait]
-impl ConsumerHandler for BatchScopeRecorder {
-    async fn handle_batch(
-        &self,
-        batch: &EventBatch<'_>,
-        _attempts: u16,
-    ) -> Result<BatchHandlerOutcome, ConsumerError> {
-        let chunk = batch.next_chunk(batch.len());
-        if let Some(first) = chunk.first() {
-            if chunk
-                .iter()
-                .any(|event| event.topic != first.topic || event.partition != first.partition)
-            {
-                self.violations
-                    .lock()
-                    .unwrap()
-                    .push("batch mixed topic IDs or partitions".to_owned());
-            }
-            self.scopes
-                .lock()
-                .unwrap()
-                .push((first.topic.clone(), first.partition, chunk.len()));
-        }
         Ok(chunk
             .last()
             .map(|event| BatchHandlerOutcome::AdvanceThrough {
